requirements.txt
langchain
langchain-community
langchain-core
langgraph
dashscope
psycopg2-binary
python-dotenv
streamlit
streamlit-agraph
pandas
sentence-transformers